+++ easy-ui-react/src/Toggle/Toggle.module.scss
@@ -1,24 +1,12 @@
 @use "../styles/common" as *;
 
 .Toggle {
-  @include component-token(
-    "toggle",
-    "color.track",
-    theme-token("color.background.gray")
-  );
-  @include component-token(
-    "toggle",
-    "color.thumb",
-    theme-token("color.text.primary")
-  );
-
   display: inline-flex;
   align-items: center;
   gap: design-token("space.1.5");
 }
 
 .standalone {
-  gap: normal;
   position: relative;
   z-index: 0;
 }
@@ -38,46 +26,4 @@
   height: 100%;
   opacity: 0;
   z-index: 1;
-}
-
-.switch {
-  display: inline-flex;
-  border-radius: 99px;
-  .focusVisible & {
-    @include native-focus-ring;
-  }
-}
-
-.track {
-  fill: component-token("toggle", "color.track");
-}
-
-.thumb {
-  fill: component-token("toggle", "color.thumb");
-}
-
-.selected {
-  @include component-token(
-    "toggle",
-    "color.track",
-    theme-token("color.background.primary")
-  );
-  @include component-token(
-    "toggle",
-    "color.thumb",
-    theme-token("color.text.primary.inverse.bold")
-  );
-}
-
-.disabled {
-  @include component-token(
-    "toggle",
-    "color.track",
-    theme-token("color.background.disabled.subtler")
-  );
-  @include component-token(
-    "toggle",
-    "color.thumb",
-    theme-token("color.text.disabled")
-  );
 }
