+++ easy-ui-react/src/Toggle/Toggle.tsx
@@ -11,6 +11,7 @@ import { useToggleState } from "react-stately";
 import { Text } from "../Text";
 import { Noop } from "../utilities/Noop";
 import { classNames } from "../utilities/css";
+import { Switch } from "./Switch";
 
 import styles from "./Toggle.module.scss";
 
@@ -57,26 +58,18 @@ export type ToggleProps = AriaLabelingProps & {
 };
 
 export function Toggle(props: ToggleProps) {
-  const { children, isDisabled, isReadOnly } = props;
+  const { children, isDisabled } = props;
 
   const ref = React.useRef(null);
   const state = useToggleState(props);
   const { inputProps: inputPropsFromSwitch } = useSwitch(props, state, ref);
   const { isFocusVisible, focusProps } = useFocusRing();
-  const { isHovered, hoverProps } = useHover(props);
+  const { hoverProps } = useHover(props);
   const isSelected = state.isSelected;
 
-  const className = classNames(
-    styles.Toggle,
-    isSelected && styles.selected,
-    isDisabled && styles.disabled,
-    isReadOnly && styles.readOnly,
-    isFocusVisible && styles.focusVisible,
-    isHovered && styles.hovered,
-    !children && styles.standalone,
-  );
-
+  const className = classNames(styles.Toggle, !children && styles.standalone);
   const textColor = isDisabled ? "disabled" : "primary";
+
   const RootComponent = children ? "label" : "span";
   const rootProps = children ? hoverProps : {};
   const InputWrapperComponent = children ? VisuallyHidden : Noop;
@@ -89,24 +82,11 @@ export function Toggle(props: ToggleProps) {
       <InputWrapperComponent>
         <input {...inputProps} className={styles.input} ref={ref} />
       </InputWrapperComponent>
-      <span className={styles.switch}>
-        <svg width={32} height={16} aria-hidden="true">
-          <rect
-            className={styles.track}
-            x={0}
-            y={0}
-            width={32}
-            height={16}
-            rx={8}
-          />
-          <circle
-            className={styles.thumb}
-            cx={isSelected ? 32 - 8 : 8}
-            cy={8}
-            r={6}
-          />
-        </svg>
-      </span>
+      <Switch
+        isSelected={isSelected}
+        isDisabled={isDisabled}
+        isFocusVisible={isFocusVisible}
+      />
       {children && (
         <span className={styles.text}>
           <Text variant="body1" color={textColor}>
